% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PropEdgeUG2D.R
\name{IndUG.PEbas.tri}
\alias{IndUG.PEbas.tri}
\title{The indicator for the presence of an edge from a point to another for the underlying or reflexivity graphs based on
Proportional Edge Proximity Catch Digraphs (PE-PCDs) - standard basic triangle case}
\usage{
IndUG.PEbas.tri(
  p1,
  p2,
  r,
  c1,
  c2,
  M = c(1, 1, 1),
  rv = NULL,
  under.graph = "OR"
)
}
\arguments{
\item{p1}{A 2D point whose PE proximity region is constructed.}

\item{p2}{A 2D point. The function determines whether there is an edge from \code{p1} to \code{p1} or not
in the underlying and reflexivity graphs of PE-PCDs.}

\item{r}{A positive real number which serves as the expansion parameter in PE proximity region; must be \eqn{\ge 1}}

\item{c1, c2}{Positive real numbers which constitute the vertex of the standard basic triangle
adjacent to the shorter edges; \eqn{c_1} must be in \eqn{[0,1/2]}, \eqn{c_2>0} and \eqn{(1-c_1)^2+c_2^2 \le 1}.}

\item{M}{A 2D point in Cartesian coordinates or a 3D point in barycentric coordinates
which serves as a center in the interior of the standard basic triangle or circumcenter of \eqn{T_b}
which may be entered as "CC" as well;
default is \eqn{M=(1,1,1)} i.e., the center of mass of \eqn{T_b}.}

\item{rv}{The index of the vertex region in \eqn{T_b} containing the point, either \code{1,2,3} or \code{NULL}
(default is \code{NULL}).}

\item{under.graph}{The type of the graph based on PE-PCDs, "OR" is for the underlying graph, and "AND" is for
the reflexivity graph (default is \code{OR}).}
}
\value{
\eqn{I(}\code{p2} is in \eqn{N_{PE}(p1,r))} for points \code{p1} and \code{p2}, that is, returns 1 if \code{p2} is in \eqn{N_{PE}(p1,r)},
returns 0 otherwise
}
\description{
Returns \eqn{I(}\code{p1p2} is an edge in the underlying or reflexivity graph of PE-PCDs \eqn{)}
for points \code{p1} and \code{p2}.

More specifically, it returns
\eqn{\max I(}\code{p2} is in \eqn{N_{PE}(p1,r))} \eqn{, I(}\code{p1} is in \eqn{N_{PE}(p2,r))} for points \code{p1} and \code{p2},
for the PE-PCD underlying graph.
That is, returns 1 if \code{p2} is in \eqn{N_{PE}(p1,r)} or \code{p1} is in \eqn{N_{PE}(p2,r)},
returns 0 otherwise,
or
\eqn{I(}\code{p2} is in \eqn{N_{PE}(p1,r))} \eqn{\times I(}\code{p1} is in \eqn{N_{PE}(p2,r))} for points \code{p1} and \code{p2},
for the PE-PCD reflexivity graph.
That is, returns 1 if \code{p2} is in \eqn{N_{PE}(p1,r)} and \code{p1} is in \eqn{N_{PE}(p2,r)},
returns 0 otherwise, where \eqn{N_{PE}(x,r)} is the PE proximity region for point \eqn{x} with expansion parameter \eqn{r \ge 1}.

PE proximity region is defined with respect to the standard basic triangle \eqn{T_b=T((0,0),(1,0),(c_1,c_2))}
where \eqn{c_1} is in \eqn{[0,1/2]}, \eqn{c_2>0} and \eqn{(1-c_1)^2+c_2^2 \le 1}.

Vertex regions are based on the center, \eqn{M=(m_1,m_2)} in Cartesian coordinates or \eqn{M=(\alpha,\beta,\gamma)} in
barycentric coordinates in the interior of the standard basic triangle \eqn{T_b} or based on circumcenter of \eqn{T_b};
default is \eqn{M=(1,1,1)} i.e., the center of mass of \eqn{T_b}.
\code{rv} is the index of the vertex region \code{p1} resides, with default=\code{NULL}.

If \code{p1} and \code{p2} are distinct and either of them are outside \eqn{T_b}, it returns 0,
but if they are identical, then it returns 1 regardless of their locations (i.e., it allows loops).

Any given triangle can be mapped to the standard basic triangle
by a combination of rigid body motions (i.e., translation, rotation and reflection) and scaling,
preserving uniformity of the points in the original triangle. Hence standard basic triangle is useful for simulation
studies under the uniformity hypothesis.

See also (\insertCite{ceyhan:Phd-thesis,ceyhan:comp-geo-2010,ceyhan:stamet2016;textual}{pcds}).
}
\examples{
\dontrun{
c1<-.4; c2<-.6
A<-c(0,0); B<-c(1,0); C<-c(c1,c2);
Tb<-rbind(A,B,C);

M<-as.numeric(pcds::runif.bas.tri(1,c1,c2)$g)

r<-1.5

P1<-as.numeric(pcds::runif.bas.tri(1,c1,c2)$g)
P2<-as.numeric(pcds::runif.bas.tri(1,c1,c2)$g)
IndUG.PEbas.tri(P1,P2,r,c1,c2,M)
IndUG.PEbas.tri(P1,P2,r,c1,c2,M,under.graph = "AND")

P1<-c(.4,.2)
P2<-c(.5,.26)
IndUG.PEbas.tri(P1,P2,r,c1,c2,M)
IndUG.PEbas.tri(P2,P1,r,c1,c2,M)

#or try
Rv<-pcds::rv.bas.tri.cent(P1,c1,c2,M)$rv
IndUG.PEbas.tri(P1,P2,r,c1,c2,M,Rv,under.graph = "AND")
}

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{IndNPEtri}} and \code{\link{IndNPETe}}
}
\author{
Elvan Ceyhan
}
