% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CentSimUG2D.R
\name{plotCSedges}
\alias{plotCSedges}
\title{The plot of the edges of the underlying or reflexivity graphs of
the Central Similarity Proximity Catch Digraph
(CS-PCD) for 2D data - multiple triangle case}
\usage{
plotCSedges(
  Xp,
  Yp,
  t,
  M = c(1, 1, 1),
  ugraph = c("underlying", "reflexivity"),
  asp = NA,
  main = NULL,
  xlab = NULL,
  ylab = NULL,
  xlim = NULL,
  ylim = NULL,
  ...
)
}
\arguments{
\item{Xp}{A set of 2D points
which constitute the vertices of the underlying
or reflexivity graphs of the CS-PCD.}

\item{Yp}{A set of 2D points
which constitute the vertices of the Delaunay triangles.}

\item{t}{A positive real number
which serves as the expansion parameter in CS proximity region.}

\item{M}{A 3D point in barycentric coordinates
which serves as a center in the interior of each Delaunay triangle,
default for \eqn{M=(1,1,1)}
which is the center of mass of each triangle.}

\item{ugraph}{The type of the graph based on CS-PCDs,
\code{"underlying"} is for the underlying graph, and \code{"reflexivity"} is for
the reflexivity graph (default is \code{"underlying"}).}

\item{asp}{A \code{numeric} value,
giving the aspect ratio \eqn{y/x} (default is \code{NA}),
see the official help page for \code{asp} by typing "\code{? asp}".}

\item{main}{An overall title for the plot (default=\code{NULL}).}

\item{xlab, ylab}{Titles for the \eqn{x} and \eqn{y} axes,
respectively (default=\code{NULL} for both).}

\item{xlim, ylim}{Two \code{numeric} vectors of length 2,
giving the \eqn{x}- and \eqn{y}-coordinate ranges
(default=\code{NULL} for both).}

\item{\dots}{Additional \code{plot} parameters.}
}
\value{
A plot of the edges of the underlying
or reflexivity graphs of the CS-PCD
whose vertices are the points in data set \code{Xp} and the Delaunay
triangles based on \code{Yp} points
}
\description{
Plots the edges of the underlying or reflexivity graphs of
the Central Similarity Proximity Catch Digraph
(CS-PCD) whose vertices are the data
points in \code{Xp} in the multiple triangle case
and the Delaunay triangles based on \code{Yp} points.

CS proximity regions are constructed
with respect to the Delaunay triangles based on \code{Yp} points, i.e.,
CS proximity regions are defined only for \code{Xp} points
inside the convex hull of \code{Yp} points.
That is, edges may exist for \code{Xp} points
only inside the convex hull of \code{Yp} points.

Edge regions in each triangle are
based on the center \eqn{M=(\alpha,\beta,\gamma)}
in barycentric coordinates in the interior of each Delaunay triangle
(default for \eqn{M=(1,1,1)}
which is the center of mass of the triangle).

Convex hull of \code{Yp} is partitioned by
the Delaunay triangles based on \code{Yp} points
(i.e., multiple triangles are the set of these Delaunay triangles
whose union constitutes the
convex hull of \code{Yp} points).
Loops are not allowed so edges are only possible
for points inside the convex hull of \code{Yp} points.

See (\insertCite{ceyhan:Phd-thesis,ceyhan:stamet2016;textual}{pcds.ugraph})
for more on the CS-PCDs.
Also, see (\insertCite{okabe:2000,ceyhan:comp-geo-2010,sinclair:2016;textual}{pcds.ugraph})
for more on Delaunay triangulation and the corresponding algorithm.
}
\examples{
\dontrun{
#nx is number of X points (target) and ny is number of Y points (nontarget)
nx<-20; ny<-5;  #try also nx<-40; ny<-10 or nx<-1000; ny<-10;

set.seed(1)
Xp<-cbind(runif(nx,0,1),runif(nx,0,1))
Yp<-cbind(runif(ny,0,.25),
runif(ny,0,.25))+cbind(c(0,0,0.5,1,1),c(0,1,.5,0,1))
#try also Yp<-cbind(runif(ny,0,1),runif(ny,0,1))

M<-c(1,1,1)  #try also M<-c(1,2,3)

t<-1.5  #try also t<-2

plotCSedges(Xp,Yp,t,M,xlab="",ylab="")
plotCSedges(Xp,Yp,t,M,xlab="",ylab="",ugraph="r")
}

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{plotCSedges.tri}}, \code{\link{plotASedges}},
\code{\link{plotPEedges}}, and \code{\link[pcds]{plotCSarcs}}
}
\author{
Elvan Ceyhan
}
