% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UndClassFunctions.R
\name{print.summary.NumEdges}
\alias{print.summary.NumEdges}
\title{Print a summary of a \code{NumEdges} \code{object}}
\usage{
\method{print}{summary.NumEdges}(x, ...)
}
\arguments{
\item{x}{An \code{object} of class \code{"summary.NumEdges"},
generated by \code{summary.NumEdges}.}

\item{\dots}{Additional parameters for \code{print}.}
}
\value{
None
}
\description{
Prints some information about the \code{object}.
}
\seealso{
\code{\link{print.NumEdges}}, \code{\link{summary.NumEdges}},
and \code{\link{plot.NumEdges}}
}
