% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UndClassFunctions.R
\name{print.summary.UndPCDs}
\alias{print.summary.UndPCDs}
\title{Print a summary of an \code{UndPCDs} \code{object}}
\usage{
\method{print}{summary.UndPCDs}(x, ...)
}
\arguments{
\item{x}{An \code{object} of class \code{"summary.UndPCDs"},
generated by \code{summary.UndPCDs}.}

\item{\dots}{Additional parameters for \code{print}.}
}
\value{
None
}
\description{
Prints some information about the \code{object}.
}
\seealso{
\code{\link{print.UndPCDs}}, \code{\link{summary.UndPCDs}},
and \code{\link{plot.UndPCDs}}
}
